Day 3/RAG_QA_System_Gemini/requirements.txt
torch
langchain
langchain-community
faiss-cpu
google-generativeai
sentence-transformers
streamlit
pypdf
python-dotenv
langchain-google-genai
numpy==1.26.4
